feat: add Block and FunctionCallBlock types
