+++ packages/ts-generator/src/types/index.ts
@@ -1,0 +1,3 @@
+export * from "./blocks";
+export * from "./modules";
+export * from "./common";
