+++ packages/ts-generator/src/types/modules.ts
@@ -1,0 +1,14 @@
+import { FunctionInfo, TypeInfo } from "types";
+
+export interface CodebaseInfo {
+  functions: FunctionInfo[];
+  types: TypeInfo[];
+}
+
+export interface ModuleInfoFields {
+  fileName: string;
+  interfaces: string[];
+  functions: FunctionInfo[];
+  fileContent: string;
+  usedTypes: string[];
+}
